backend/requirements.txt
# ScrapIt - Consolidated Requirements

# Core Framework
fastapi>=0.110.0
uvicorn>=0.29.0
sqlalchemy>=2.0.27
psycopg2-binary>=2.9.9

# Authentication
PyJWT>=2.8.0
google-api-python-client>=2.120.0
google-auth>=2.40.0
google-auth-oauthlib>=1.2.0

# AI
openai>=1.20.0

# Utilities
python-dotenv>=1.0.0
cryptography>=42.0.0